+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Un-Released]
 - Upgrade mina-signer to 3.0.0
 - Update some i18n 
+- Add Russian language support
 
 ## [2.2.12]
 - Fix account init